+++ src/controllers/authController.ts
@@ -1,7 +1,5 @@
 import { Request, Response, NextFunction } from "express";
 import bcrypt from "bcrypt";
-import { UPLOAD_DIR } from '../app';
-import path from "path";
 import jwt, { JwtPayload } from "jsonwebtoken";
 import logger from "../config/logger";
 
@@ -33,7 +31,7 @@ import { actualizarFechaFinSesion, registrarAuditoria, registrarSesion } from ".
 
 import * as Err from "../utils/customErrors";
 import * as MESSAGES from "../utils/messages";
-import { ProductoraDocumento, Usuario } from "../models";
+import { ProductoraDocumento } from "../models";
 
 
 // LOGIN
@@ -140,7 +138,7 @@ export const login = async (
 
     res.cookie("auth_token", token, {
       httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
+      secure: process.env.NODE_ENV?.startsWith("production"),
       sameSite: "strict",
       maxAge: parseInt(process.env.COOKIE_MAX_AGE || "3600000", 10),
     });
@@ -193,7 +191,7 @@ export const logout = async (
     // Limpiar la cookie de autenticación
     res.clearCookie("auth_token", {
       httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
+      secure: process.env.NODE_ENV?.startsWith("production"),
       sameSite: "strict",
     });
 
@@ -201,7 +199,7 @@ export const logout = async (
     if (activeSesion) {
       res.clearCookie("active_sesion", {
         httpOnly: true,
-        secure: process.env.NODE_ENV === "production",
+        secure: process.env.NODE_ENV?.startsWith("production"),
         sameSite: "strict",
       });
 
@@ -550,7 +548,7 @@ export const selectAuthProductora = async (
       }),
       {
         httpOnly: true,
-        secure: process.env.NODE_ENV === "production",
+        secure: process.env.NODE_ENV?.startsWith("production"),
         sameSite: "strict",
         maxAge: parseInt(process.env.COOKIE_MAX_AGE || "3600000", 10),
       }
@@ -794,7 +792,7 @@ export const resetPassword = async (
     // Limpia la cookie de autenticación para cerrar sesión
     res.clearCookie("auth_token", {
       httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
+      secure: process.env.NODE_ENV?.startsWith("production"),
       sameSite: "strict",
     });